feat(router): router restructure
