+++ src/app/shared/components/index.tsx
@@ -1,2 +1,3 @@
 export * from './layout';
 export * from './header';
+export * from './routing';
