+++ src/app/shared/components/routing/index.tsx
@@ -1,0 +1,1 @@
+export * from './routing.component';
